codegen: share RecursionPlan enum + AST transform between backends
Pull the recursion classifier's `RecursionPlan` / `ProofModeIssue`
types plus the `rewrite_recursive_calls_{expr,body}` transform into a
new `codegen::recursion` module. Lean re-exports the enum and delegates
to the shared transform; the Dafny backend will pick the same shared
API in the next step.

Also flips `fuel_helper_name` in Lean to the shared helper so the name
emitted into `Expr::Ident("foo__fuel")` matches the name on the helper
definition — previously Lean locally wrapped the original name in
`aver_name_to_lean` before the suffix, which would have diverged from
the shared rewrite if it ever emitted its own call site.

Behaviour is unchanged: 470 lib + 35 proof-spec tests stay green.
