+++ src/codegen/lean/mod.rs
@@ -34,43 +34,11 @@ pub enum VerifyEmitMode {
     TheoremSkeleton,
 }
 
-/// Proof-mode recursion support class for emitted Lean definitions.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum RecursionPlan {
-    /// Single-function recursion where an `Int` parameter decreases by 1.
-    IntCountdown { param_index: usize },
-    /// Single-function recursion where an `Int` parameter increases by 1 up to a bound.
-    /// `bound_lean` is the Lean expression of the upper bound (e.g. `8` or `xs.length`).
-    IntAscending {
-        param_index: usize,
-        bound_lean: String,
-    },
-    /// Single-function recurrence with `n < 0` guard, `0/1` bases, and a linear
-    /// second-order recurrence over previous results, emitted through a private
-    /// Nat helper.
-    LinearRecurrence2,
-    /// Single-function structural recursion on a `List<_>` parameter.
-    ListStructural { param_index: usize },
-    /// Single-function structural recursion on a recursive user-defined type parameter.
-    SizeOfStructural,
-    /// Single-function recursion where first `String` stays the same and second `Int`
-    /// parameter strictly advances (`pos + k`, `k >= 1`).
-    StringPosAdvance,
-    /// Mutual recursion group where first `Int` parameter decreases by 1 across SCC calls.
-    MutualIntCountdown,
-    /// Mutual recursion group where first `String` is preserved and second `Int`
-    /// either stays the same across rank-decreasing edges, or strictly advances.
-    MutualStringPosAdvance { rank: usize },
-    /// Generic mutual recursion plan using `sizeOf` on structural parameters,
-    /// plus rank for same-measure edges.
-    MutualSizeOfRanked { rank: usize },
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ProofModeIssue {
-    pub line: usize,
-    pub message: String,
-}
+// RecursionPlan / ProofModeIssue moved to shared `crate::codegen::recursion`
+// so the Dafny backend can reuse the same classifier. Re-export here so
+// existing `lean::RecursionPlan` / `lean::ProofModeIssue` call sites keep
+// working without churn.
+pub use crate::codegen::recursion::{ProofModeIssue, RecursionPlan};
 
 const LEAN_PRELUDE_HEADER: &str = r#"-- Generated by the Aver → Lean 4 transpiler
 -- Pure core logic plus Oracle-lifted classified effects
